added spreate page for student and teacher
