--- lib/Pages/Leave/LeaveForm.dart
@@ -1,49 +1,0 @@
-import 'package:flutter/material.dart';
-import 'package:flutter/widgets.dart';
-import 'package:flutter_svg/flutter_svg.dart';
-import 'package:get/get.dart';
-import 'package:myapp/Config/images.dart';
-import 'package:myapp/Pages/Leave/Widgets/LeaveSheetBox.dart';
-
-class LeaveForm extends StatefulWidget {
-  @override
-  _LeaveFormState createState() => _LeaveFormState();
-}
-
-class _LeaveFormState extends State<LeaveForm> {
-  @override
-  Widget build(BuildContext context) {
-    return Scaffold(
-      backgroundColor: Colors.white,
-      appBar: AppBar(
-        backgroundColor: Colors.white,
-        centerTitle: true,
-        title: Text(
-          "Leave Sheet",
-          style: Theme.of(context).textTheme.displayMedium,
-        ),
-        leading: InkWell(
-          onTap: () {
-            Get.back();
-          },
-          child: const Icon(Icons.arrow_back_ios_new_outlined),
-        ),
-      ),
-      body: Stack(
-        children: [
-          SvgPicture.asset(
-            AssetsIamge.bgDesignSVG,
-            fit: BoxFit.cover,
-            width: double.infinity,
-            height: double.infinity,
-          ),
-          ListView(
-            children: [
-              LeaveSheetBox()
-            ],
-          ),
-        ],
-      ),
-    );
-  }
-}
