+++ lib/Pages/Student_Page/ask_doubt/askDoubt.dart
@@ -1,0 +1,53 @@
+import 'package:flutter/material.dart';
+import 'package:flutter_svg/svg.dart';
+import 'package:get/get.dart';
+import 'package:myapp/Config/images.dart';
+import 'package:myapp/Pages/Student_Page/ask_doubt/widgets/askDoubtBox.dart';
+
+class AskDoubtPage extends StatefulWidget {
+  const AskDoubtPage({super.key});
+
+  @override
+  State<AskDoubtPage> createState() => _AskDoubtPageState();
+}
+
+class _AskDoubtPageState extends State<AskDoubtPage> {
+  @override
+  Widget build(BuildContext context) {
+    return Scaffold(
+      backgroundColor: Colors.white,
+      appBar: AppBar(
+        backgroundColor: Colors.white,
+        centerTitle: true,
+        title: Text(
+          "Ask Doubt",
+          style: Theme.of(context).textTheme.displayMedium,
+        ),
+        leading: InkWell(
+          onTap: () {
+            Get.back();
+          },
+          child: const Icon(Icons.arrow_back_ios_new_outlined),
+        ),
+      ),
+      body: Stack(
+        children: [
+          SvgPicture.asset(
+            AssetsIamge.bgDesignSVG,
+            fit: BoxFit.cover,
+            width: double.infinity,
+            height: double.infinity,
+          ),
+          Padding(
+            padding: const EdgeInsets.only(right: 20, left: 20),
+            child: ListView(
+              children: [
+                AskDoubtBox()
+              ],
+            ),
+          )
+        ],
+      ),
+    );
+  }
+}
